+++ howdju-service-common/lib/extractQuotation.test.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from "fs";
 import { JSDOM } from "jsdom";
 import stripIndent from "strip-indent";
-import { extractQuotationFromWindowsFragmentUsingPolyfill } from "./extractQuotation";
+import { extractQuotationFromWindowFragmentUsingPolyfill } from "./extractQuotation";
 
 describe("extractQuotationFromWindowsFragmentUsingPolyfill", () => {
   // TODO(506) speed up this test and then maybe use
@@ -15,7 +15,7 @@ describe("extractQuotationFromWindowsFragmentUsingPolyfill", () => {
       "https://lexfridman.com/robert-f-kennedy-jr-transcript/#:~:text=Camus-,Lex%20Fridman,act%20of%20rebellion.%E2%80%9D%20What%20do%20you%20think%20he%20means%20by%20that%3F,-Robert%20F.%20Kennedy";
     const dom = new JSDOM(html, { url, runScripts: "outside-only" });
 
-    expect(extractQuotationFromWindowsFragmentUsingPolyfill(dom.window)).toBe(
+    expect(extractQuotationFromWindowFragmentUsingPolyfill(dom.window)).toBe(
       stripIndent(`
         Lex Fridman (00:09:26) The blood of each generation. You mentioned your interest, your admiration of Al Albert Camus, of Stoicism, perhaps your interest in existentialism. Camus said, I believe in Myth of Sisyphus, “The only way to deal with an unfree world is to become so absolutely free that your very existence is an act of rebellion.” What do you think he means by that?
         `).trim()
