+++ howdju-text-fragments/src/window.d.ts
@@ -1,5 +1,8 @@
 import { GenerateFragmentResult } from "text-fragments-polyfill/dist/fragment-generation-utils.js";
 
+/**
+ * Augments the Window interface with the methods we provide in this package.
+ */
 declare global {
   interface Window {
     rangeToFragment: (range: Range) => GenerateFragmentResult;
