+++ package.json
@@ -23,6 +23,7 @@
     "fix-format": "yarn run lint --fix && yarn run prettier --write --ignore-path .gitignore jest lambdas/bin",
     "fix-format:all": "echo format:all && yarn workspaces foreach -Apv run fix-format && echo fix-format:all succeeded || {echo fix-format:all failed; exit 1}",
     "lint": "eslint jest lambdas/bin",
+    "rm:all:node_modules": "yarn workspaces foreach -Apv exec rm -r node_modules && rm -r node_modules",
     "merge:coverage": "bin/merge-coverage.sh",
     "lint:all": "echo lint:all && yarn workspaces foreach -Apv run lint && echo lint:all succeeded || {echo lint:all failed; exit 1}",
     "custom-check:all": "echo custom-check:all && yarn workspaces foreach -Apv run custom-check && echo custom-check:all succeeded || {echo custom-check:all faied; exit 1}",
